refactor: HTML 유틸 관련 분리
