+++ src/pages/Main/Main.js
@@ -10,10 +10,16 @@ import SwiperCore, {
 } from 'swiper/core';
 
 import SEO from 'components/SEO';
+import MainRecommend from 'components/Main/MainRecommend';
 import MainEvent from 'components/Main/MainEvent';
 import CustomerService from 'components/CustomerService';
 import Alert from 'components/common/Alert';
 import { useWindowSize } from 'utils/utils';
+import {
+    getRecommendedBannerNames,
+    getSlideBannerNames,
+    getAcademyBannerNames,
+} from 'utils/html';
 import { breakPoint } from 'utils/constants';
 import { main } from 'const/seo';
 import { getDisplaySectionsSectionNo, loadBanner } from 'api/display';
@@ -24,7 +30,6 @@ import 'swiper/components/pagination/pagination.scss';
 import 'swiper/components/scrollbar/scrollbar.scss';
 import 'swiper/swiper.scss';
 import 'assets/scss/main.scss';
-import MainRecommend from 'components/Main/MainRecommend';
 
 export default function Main() {
     const { openAlert, closeModal, alertVisible, alertMessage } = useAlert();
@@ -53,57 +58,6 @@ export default function Main() {
     const [recommendedSections, setRecommendedSections] = useState([]);
 
     const [eventSections, setEventSections] = useState();
-
-    const getRecommendedBannerNames = (bannerInfoList) => {
-        bannerInfoList.forEach((bannerInfo) => {
-            const bannerNameList = bannerInfo.banners[0].name.split('/');
-            bannerInfo.banners[0].nameList = bannerNameList.reduce(
-                (acc, bannerName, index) => {
-                    if (bannerNameList.length - 1 === index) {
-                        acc += `${bannerName}`;
-                    } else {
-                        acc += `${bannerName}<br />`;
-                    }
-                    return acc;
-                },
-                '',
-            );
-        });
-    };
-
-    const getSlideBannerNames = (bannerInfoList) => {
-        bannerInfoList.forEach((bannerInfo) => {
-            let bannerNameList = bannerInfo.banners[0].name.split('/');
-            bannerNameList = bannerNameList.map((name) => name.split(' '));
-            let count = 0;
-            bannerInfo.banners[0].nameList = bannerNameList.reduce(
-                (acc, bannerName) => {
-                    const nameHtml = bannerName.reduce((acc, name) => {
-                        acc += `<span class="copy-${count}"><span>${name}</span></span>`;
-                        count++;
-                        return acc;
-                    }, '');
-                    acc += `<div class="kv__head__copy">${nameHtml}</div>`;
-                    return acc;
-                },
-                '',
-            );
-        });
-    };
-
-    const getAcademyBannerNames = (bannerInfoList) => {
-        if (Object.keys(bannerInfoList).length === 0) return;
-        const bannerNames = bannerInfoList.banners[0].name.split('/');
-        bannerInfoList.banners[0].nameList = bannerNames.reduce(
-            (acc, bannerName, index) => {
-                const { length } = bannerNames;
-                acc +=
-                    index - 1 === length ? bannerName : bannerName + '<br />';
-                return acc;
-            },
-            '',
-        );
-    };
 
     //1. 배너 노출 api
     const getBanners = useCallback(async () => {
